# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ФИТИТ — Персональный план питания API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /plans:
    post:
      operationId: createPlan
      tags: [plans]
      summary: Purchase and generate a new meal plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    get:
      operationId: listPlans
      tags: [plans]
      summary: List all user plans
      responses:
        "200":
          description: List of plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
  /plans/current:
    get:
      operationId: getCurrentPlan
      tags: [plans]
      summary: Get the user's active plan with all meals
      responses:
        "200":
          description: Current plan with meals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanWithMeals"
        "404":
          description: No active plan
  /plans/{planId}:
    get:
      operationId: getPlan
      tags: [plans]
      summary: Get a specific plan with all meals
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan with meals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanWithMeals"
        "404":
          description: Plan not found
  /plans/{planId}/regenerate:
    post:
      operationId: regeneratePlan
      tags: [plans]
      summary: Regenerate the entire meal plan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan regenerated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanWithMeals"
  /plans/{planId}/meals/{mealId}:
    patch:
      operationId: replaceMeal
      tags: [plans]
      summary: Replace a single meal in the plan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
        - name: mealId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealReplaceInput"
      responses:
        "200":
          description: Meal replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
  /plans/{planId}/export/{format}:
    get:
      operationId: exportPlan
      tags: [plans]
      summary: Export plan as PDF or DOCX download URL
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
        - name: format
          in: path
          required: true
          schema:
            type: string
            enum: [pdf, docx]
      responses:
        "200":
          description: Export URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponse"
  /plans/{planId}/summary:
    get:
      operationId: getPlanSummary
      tags: [plans]
      summary: Get plan macro averages and daily breakdown summary
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanSummary"
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    User:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
...
    Plan:
      type: object
      required: [id, userId, duration, startDate, endDate, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        duration:
          type: string
          enum: [week, month, three_months, six_months]
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [active, expired, cancelled]
        createdAt:
          type: string
    PlanInput:
      type: object
      required: [duration]
      properties:
        duration:
          type: string
          enum: [week, month, three_months, six_months]
    Meal:
      type: object
      required: [id, planId, dayNumber, mealNumber, mealType, name, ingredients, recipe, calories, proteins, fats, carbs]
      properties:
        id:
          type: integer
        planId:
          type: integer
        dayNumber:
          type: integer
        mealNumber:
          type: integer
        mealType:
          type: string
        name:
          type: string
        ingredients:
          type: string
        recipe:
          type: string
        calories:
          type: integer
        proteins:
          type: integer
        fats:
          type: integer
        carbs:
          type: integer
    MealReplaceInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: ["string", "null"]
    PlanWithMeals:
      type: object
      required: [id, userId, duration, startDate, endDate, status, createdAt, meals]
      properties:
        id:
          type: integer
        userId:
          type: integer
        duration:
          type: string
          enum: [week, month, three_months, six_months]
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [active, expired, cancelled]
        createdAt:
          type: string
        meals:
          type: array
          items:
            $ref: "#/components/schemas/Meal"
    PlanSummary:
      type: object
      required: [avgCalories, avgProteins, avgFats, avgCarbs, totalDays, mealsPerDay]
      properties:
        avgCalories:
          type: integer
        avgProteins:
          type: integer
        avgFats:
          type: integer
        avgCarbs:
          type: integer
        totalDays:
          type: integer
        mealsPerDay:
          type: integer
        weeklyCalories:
          type: ["integer", "null"]
    ExportResponse:
      type: object
      required: [url, format]
      properties:
        url:
          type: string
        format:
          type: string