# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ФИТИТ — Персональный план питания API
 * OpenAPI spec version: 0.1.0
 */
import type { PlanDuration } from './planDuration';
import type { PlanStatus } from './planStatus';

export interface Plan {
  id: number;
  userId: number;
  duration: PlanDuration;
  startDate: string;
  endDate: string;
  status: PlanStatus;
  createdAt: string;
}