# lib/api-zod/src/generated/types/planWithMeals.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ФИТИТ — Персональный план питания API
 * OpenAPI spec version: 0.1.0
 */
import type { Meal } from './meal';
import type { PlanWithMealsDuration } from './planWithMealsDuration';
import type { PlanWithMealsStatus } from './planWithMealsStatus';

export interface PlanWithMeals {
  id: number;
  userId: number;
  duration: PlanWithMealsDuration;
  startDate: string;
  endDate: string;
  status: PlanWithMealsStatus;
  createdAt: string;
  meals: Meal[];
}