# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ФИТИТ — Персональный план питания API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}